SCAN4NET-331 Extract ParameterTest
